\name{CP}
\alias{CP}
\alias{CP.snp}

\title{Cellularity and dna.index results sample.}

\description{
  Examples of objects resulting from the Bayesian estimation over a set of \code{cellualrity} and \code{dna.index} parameters. 
}

\examples{
  data(CP.example)
  str(CP)
  data(CPsnp.example)
  str(CP.snp)
}
