\name{CP.example}
\alias{CPsnp.example}
\alias{CP.example}
\alias{CP}
\alias{CP.snp}

\title{Example of cellularity and ploidy results.}

\description{
  Examples of objects resulting from the Bayesian estimation over a set of cellularity and ploidy parameters. 
}

\examples{
  data(CP.example)
  str(CP.example)
  data(CPsnp.example)
  str(CPsnp.example)
}
