\name{CP.example}
\alias{CP.example}
\alias{CPsnp.example}

\title{Cellularity and dna.index results sample.}

\description{
  Example of objects resulting from the Bayesian estimation over a set of \code{cellualrity} and \code{dna.index} parameters. 
}

\usage{
  data(CP.example)
  data(CPsnp.example)
}
